requirements.txt
customtkinter
pandas
matplotlib
pyvis
CTkMessagebox
pillow
numpy
pyinstaller
cairosvg
